Move transactions-modal into dash-transactions
This is done to make the component structure more similar
to dash-budgets and budgets-dialog.
Several things have also been renamed, such as CSS class names and the
component name.
