+++ src/app/app.module.ts
@@ -38,9 +38,8 @@ import { AppComponent } from './app.component';
 import { AlertComponent } from './components/alert/alert.component';
 import { DashboardComponent } from './routes/dashboard/dashboard.component';
 import { DashHomeComponent } from './components/dash-home/dash-home.component';
-import { DashTransactionsComponent } from './components/dash-transactions/dash-transactions.component';
+import { DashTransactionsComponent, TransactionsDialogComponent } from './components/dash-transactions/dash-transactions.component';
 import { PageNotFoundComponent } from './components/page-not-found/page-not-found.component';
-import { TransactionsModalComponent } from './components/transactions-modal/transactions-modal.component';
 import { BudgetsDialogComponent, DashBudgetsComponent } from './components/dash-budgets/dash-budgets.component';
 import { DashSettingsComponent } from './components/dash-settings/dash-settings.component';
 import { DashCalendarComponent } from './components/dash-calendar/dash-calendar.component';
@@ -48,6 +47,7 @@ import { CalendarModule, DateAdapter } from 'angular-calendar';
 import { adapterFactory } from 'angular-calendar/date-adapters/date-fns';
 import { DashProgressComponent } from './components/dash-progress/dash-progress.component';
 import { EventCardComponent } from './components/event-card/event-card.component';
+import { MatFormFieldModule } from '@angular/material/form-field';
 
 // Services
 
@@ -61,8 +61,8 @@ import { EventCardComponent } from './components/event-card/event-card.component
     DashboardComponent,
     DashHomeComponent,
     DashTransactionsComponent,
+    TransactionsDialogComponent,
     PageNotFoundComponent,
-    TransactionsModalComponent,
     DashBudgetsComponent,
     BudgetsDialogComponent,
     DashSettingsComponent,
@@ -93,6 +93,7 @@ import { EventCardComponent } from './components/event-card/event-card.component
     MatSelectModule,
     MatCheckboxModule,
     MatTabsModule,
+    MatFormFieldModule,
     CalendarModule.forRoot({ provide: DateAdapter, useFactory: adapterFactory })
   ],
   providers: [],
