+++ src/app/components/dash-transactions/transactions-dialog.html
@@ -1,7 +1,7 @@
-<div class="transaction-modal">
+<div class="transactions-dialog">
   <h2 mat-dialog-title>{{ mode }} Transaction</h2>
   <form
-    class="transaction-modal__form"
+    class="transactions-dialog__form"
     mat-dialog-content
     [formGroup]="transactionForm"
     (ngSubmit)="goSubmitTransaction()"
