--- src/app/components/transactions-modal/transactions-modal.component.css
@@ -1,7 +1,0 @@
-.transaction-modal__form {
-  /* self align */
-  min-width: 500px;
-  /* content align */
-  display: flex;
-  flex-direction: column;
-}
